:hammer: Move Precept to own file.
