+++ precept/__init__.py
@@ -1,5 +1,6 @@
 from ._version import __version__  # noqa: F401
-from ._cli import Precept, Command, Argument  # noqa: F401
+from ._cli import Command, Argument  # noqa: F401
+from ._precept import Precept  # noqa: F401
 from ._tools import *  # noqa: F401, F403
 from ._immutable import *  # noqa: F401, F403
 from ._configs import (  # noqa: F401
